+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - Fixed a couple of the "box" snippets that had duplicate last values.
+- Improved compiling configuration
+- Added error message if missing povray/pvengine config setting
 
 ## [2.1.0] - 2022-03-02